+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Changed
+- Use GHC.Compact to prevent excessive GC
+
 ## [v0.1.0.0] - 2019-05-02
 
 ### Added